Move the indexer into the shared library
